fix: changed amount data types to double

+++ src/database/migrations/20210624060218-create-product.js
@@ -32,7 +32,7 @@ module.exports = {
         allowNull: false
       },
       price: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.DOUBLE,
         allowNull: false
       },
       created_at: {
